+++ build_native/lib/build_native.dart
@@ -1,3 +1,4 @@
-export 'src/cmake_builder.dart';
-//export 'src/object_file_builder.dart';
-export 'src/windows.dart';
+//export 'src/cmake_builder.dart';
+export 'src/library_builder.dart';
+export 'src/object_file_builder.dart';
+//export 'src/windows.dart';