+++ build_native/lib/src/builder.dart
@@ -1,3 +1,4 @@
+/*
 import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
@@ -388,3 +389,4 @@ class NativeExtensionBuilder implements Builder {
     await process.stdout.toList();
   }
 }
+*/