+++ build_native/lib/src/compiler/macos.dart
@@ -1,0 +1,11 @@
+import 'unix.dart';
+
+class MacOSNativeExtensionCompiler extends UnixNativeExtensionCompiler {
+  const MacOSNativeExtensionCompiler();
+
+  @override
+  String get defaultCC => 'clang';
+
+  @override
+  String get defaultCXX => 'clang';
+}