+++ example/bin/main.dart
@@ -1,5 +1,5 @@
-import '../lib/sample_extension.dart';
-import '../lib/say_hello.dart';
+import 'package:example/sample_extension.dart';
+import 'package:example/say_hello.dart';
 
 main() {
   var num = systemRand();